+++ apps/docs/src/app/core/component-docs/vertical-navigation/examples/vertical-navigation-grouping-example.component.ts
@@ -1,0 +1,8 @@
+import { Component, ViewChild } from '@angular/core';
+import { ListNavigationItemComponent } from '@fundamental-ngx/core';
+
+@Component({
+    selector: 'fd-vertical-navigation-grouping-example',
+    templateUrl: './vertical-navigation-grouping-example.component.html'
+})
+export class VerticalNavigationGroupingExampleComponent {}