+++ apps/docs/src/app/core/documentation/core-documentation.component.ts
@@ -90,6 +90,7 @@ export class CoreDocumentationComponent extends DocumentationBaseComponent {
             { url: 'core/token', name: 'Token' },
             { url: 'core/toolbar', name: 'Toolbar' },
             { url: 'core/object-number', name: 'Object Number' },
+            { url: 'core/vertical-navigation', name: 'Vertical Navigation'},
             { url: 'core/upload-collection', name: 'Upload Collection' },
             { url: 'core/wizard', name: 'Wizard' },
             { url: 'core/grid-list', name: 'Grid List' },