+++ libs/core/src/lib/pagination/pagination.component.ts
@@ -16,9 +16,10 @@ import {
     TemplateRef,
     ViewEncapsulation
 } from '@angular/core';
-import '@angular/localize/init';
-import { KeyUtil, RtlService } from '@fundamental-ngx/core/utils';
 import { Subscription } from 'rxjs';
+
+import { KeyUtil, RtlService } from '@fundamental-ngx/core/utils';
+
 import { Pagination } from './pagination.model';
 import { PaginationService } from './pagination.service';
 
@@ -29,7 +30,7 @@ interface CurrentShowing {
     from: number;
     to: number;
     of: number;
-};
+}
 
 let paginationUniqueId = 0;
 
@@ -100,11 +101,9 @@ export class PaginationComponent implements OnChanges, OnInit, OnDestroy {
     /**
      * Label for options for items per page.
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Results per page`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    itemsPerPageLabel = $localize`:@@corePaginationItemsPerPageLabel:Results per page`;
+    itemsPerPageLabel = 'Results per page';
 
     /** Represents the options for items per page. */
     @Input()
@@ -137,47 +136,44 @@ export class PaginationComponent implements OnChanges, OnInit, OnDestroy {
     /**
      * Label for the 'previous' page button.
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Previous`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    previousLabel = $localize`:@@corePaginationPreviousLabel:Previous`;
+    previousLabel = 'Previous';
 
     /**
      * Label for the 'next' page button.
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Next`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    nextLabel = $localize`:@@corePaginationNextLabel:Next`;
+    nextLabel = 'Next';
 
     /**
      * Label for the 'Page' page button.
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Page`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    pageLabel = $localize`:@@corePaginationPageLabel:Page `;
+    pageLabel = 'Page';
 
     /**
      * Aria label for the navigation element
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Pagination`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    ariaLabel = $localize`:@@corePaginationAriaLabel:Pagination`;
+    ariaLabel = 'Pagination';
 
     /**
      * The current page label that should be read when a page is selected.
+     * Please use ${currentPage} in the value for replacing with the current page number.
+     * Example: `Page ${currentPage} is selected`.
      * This property is mainly provided to support reading in the right language for screen reader.
-     * Here, Angular i18n's `$localize` tag is used with a default text of `Current page`.
-     * Application developer is expected to pass in their own value if they use any other localization tool.
      */
     @Input()
-    currentPageAriaLabel = $localize`:@@corePaginationCurrentPageAriaLabel: Page ${this.currentPage}:page: is current page`;
+    currentPageAriaLabel = 'Page ${currentPage} is current page';
+
+    /** @hidden */
+    get _currentPageAriaLabel(): string {
+        return this.currentPageAriaLabel.replace(/\${currentPage}/, this.currentPage.toString());
+    }
 
     /** Event fired when the page is changed. */
     @Output()
@@ -299,14 +295,16 @@ export class PaginationComponent implements OnChanges, OnInit, OnDestroy {
         if ($event) {
             $event.preventDefault();
         }
+
         if (page > this.paginationService.getTotalPages(this.getPaginationObject()) || page < 1) {
             return;
         }
+
         this._refreshPages();
 
         this.pageChangeStart.emit(page);
 
-        this._liveAnnouncer.announce(this.currentPageAriaLabel);
+        this._liveAnnouncer.announce(this._currentPageAriaLabel);
     }
 
     /**