+++ libs/core/src/lib/public_api.ts
@@ -83,6 +83,7 @@ export * from '@fundamental-ngx/core/status-indicator';
 export * from '@fundamental-ngx/core/shared';
 export * from '@fundamental-ngx/core/mobile-mode';
 export * from '@fundamental-ngx/core/resizable-card-layout';
+export * from '@fundamental-ngx/core/vertical-navigation';
 export * from '@fundamental-ngx/core/upload-collection';
 export * from '@fundamental-ngx/core/micro-process-flow';
 export * from '@fundamental-ngx/core/timeline';